backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pymongo<4.6
motor==3.3.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
face-recognition==1.3.0
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models
reportlab==4.2.2
